Fixed file name association for better youtube video searching
I did not include Nikko's algorithm yet because it's still quite
unstable. He seems to be having trouble getting the indices of the
mistakes and there are still instances where it crashes because of it,
so I will include it in the next commit if we're able to fix it.